+++ api/app.py
@@ -19,21 +19,6 @@ along with this program.  If not, see <https://www.gnu.org/licenses/>.
 from fastapi import FastAPI
 
 from api.schema import graphql_app
-from database.engine import init_db
 
 app = FastAPI()
 app.include_router(graphql_app, prefix='/graphql')
-
-
-@app.on_event('startup')
-async def on_startup() -> None:
-    """
-    Create database.
-
-    Todo:
-    ----
-    - Move from events to lifespan
-    see https://fastapi.tiangolo.com/advanced/events/
-
-    """
-    await init_db()