+++ api/resolvers.py
@@ -18,7 +18,7 @@ along with this program.  If not, see <https://www.gnu.org/licenses/>.
 
 from pydantic import PositiveInt
 
-from api.address.inputs import AddressFilterInput, AddressInsertInput
+from api.address.graphql_inputs import AddressFilterInput, AddressInsertInput
 from database import functions
 from database.models.brazil import Address
 from plugins.plugins_controller import get_zipcode_from_plugins
@@ -57,4 +57,18 @@ async def get_address(
 
 
 async def insert_address(address: AddressInsertInput) -> Address:
+    """
+    Insert address and city if not exists on database.
+
+    Parameters
+    ----------
+    address : AddressInsertInput
+        Strict address class, all needed fields need to be passed
+
+    Returns
+    -------
+    AddressType
+        Address (db model)
+
+    """
     return await functions.insert_address_by_dc(address)